Move xivfixes.h to separate ifle
